+++ server/entidades/Discapacidad.php
@@ -6,7 +6,7 @@ class Discapacidad
    public $idTipoDiscapacidad;
    public $porcentaje;
 
-   function __construct(int $id,int $idPersona,string $idTipoDiscapacidad,double $porcentaje){
+   function __construct(int $id,int $idPersona,string $idTipoDiscapacidad,float $porcentaje){
       $this->id = $id;
       $this->idPersona = $idPersona;
       $this->idTipoDiscapacidad = $idTipoDiscapacidad;
